requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydantic==2.7.1